% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhhash.R
\name{build_hash_from_response}
\alias{build_hash_from_response}
\title{Build a hash from headers in an {curl} response object}
\usage{
build_hash_from_response(response, version = 1L)
}
\arguments{
\item{response}{an {curl} response object}

\item{version}{hhhash version (defaults to \code{1L})}
}
\value{
A hash of headers
}
\description{
Build a hash from headers in an {curl} response object
}
\examples{
res <- curl::curl_fetch_memory("https://www.circl.lu/", curl::new_handle())

build_hash_from_response(res)
}
