% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhhash.R
\name{build_hash_from_url}
\alias{build_hash_from_url}
\title{Build a hash from headers retrieved from a URL}
\usage{
build_hash_from_url(url, handle = curl::new_handle(), version = 1L)
}
\arguments{
\item{url}{URL to retrieve headers from}

\item{handle}{a {curl} handle; by default it creates a new one}

\item{version}{hhhash version (defaults to \code{1L})}
}
\value{
A hash of headers
}
\description{
Build a hash from headers retrieved from a URL
}
\examples{
build_hash_from_url("https://www.circl.lu/")
}
