% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhhash.R
\name{hash_headers}
\alias{hash_headers}
\title{Build a hash from a vector of HTTP header keys}
\usage{
hash_headers(keys, version = 1L)
}
\arguments{
\item{keys}{a character vector of HTTP header keys}

\item{version}{hhhash version (defaults to \code{1L})}
}
\value{
A hash of headers
}
\description{
Build a hash from a vector of HTTP header keys
}
\examples{
c(
  "Date", "Server", "Strict-Transport-Security",
  "Last-Modified", "ETag", "Accept-Ranges",
  "Content-Length", "Content-Security-Policy",
  "X-Content-Type-Options", "X-Frame-Options",
  "X-XSS-Protection", "Content-Type"
) -> keys

hash_headers(keys)
}
