% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhhash-package.R
\docType{package}
\name{hhhash}
\alias{hhhash}
\alias{hhhash-package}
\title{HTTP Headers Hashing}
\description{
HTTP Headers Hashing (HHHash) is a technique used to create a
fingerprint of an HTTP server based on the headers it returns.
HHHash employs one-way hashing to generate a hash value for the
set of header keys returned by the server
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/hhhash}
  \item Report bugs at \url{https://github/hrbrmstr/hhhash/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
