% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sha256}
\alias{sha256}
\title{Convience wrapper for digest::digest() with algo  = "sha256" and serialize = FALSE}
\usage{
sha256(obj)
}
\arguments{
\item{obj}{object to make a hash for (ideally a length 1 character vector)}
}
\description{
Convience wrapper for digest::digest() with algo  = "sha256" and serialize = FALSE
}
